Dodany skrypt do okna modalnego z teamem
i drobne poprawki w stylowaniu okna
