+++ css/main.css
@@ -367,9 +367,9 @@ a {
   background-color: rgba(0, 0, 0, 0.5);
   transition: all 250ms ease;
 }
-.modal-team__backdrop.is-hidden {
+
+.team-is-hidden {
   opacity: 0;
-  visibility: hidden;
   pointer-events: none;
 }
 
@@ -385,7 +385,7 @@ a {
 .modal-team__members {
   position: relative;
   margin: 0 auto;
-  padding: 10px;
+  padding: 25px;
   width: 100%;
   background-color: #f5f4fa;
 }
@@ -393,6 +393,7 @@ a {
   font-size: 20px;
   line-height: 1.25;
   text-align: center;
+  margin-bottom: 15px;
 }
 .modal-team__list {
   display: flex;
