+++ js/team-modal.js
@@ -1,0 +1,14 @@
+const qs = (sel) => document.querySelector(sel);
+
+const openModal = qs("[data-open-team]");
+const closeModal = qs("[data-close-team]");
+const modal = qs("[data-backdrop]");
+console.log(openModal);
+console.log(closeModal);
+
+const toggleModal = () => {
+  modal.classList.toggle("team-is-hidden");
+};
+
+openModal.addEventListener("click", toggleModal);
+closeModal.addEventListener("click", toggleModal);
