+++ js/trendMovies.js
@@ -1,13 +1,13 @@
-import { fetchPopularMovie } from "./fetchApi";
-import { renderMovies } from "./renderMovies";
+// import { fetchPopularMovie } from "./fetchApi";
+// import { renderMovies } from "./renderMovies";
 
-export default function setPopularMovie() {
-  fetchPopularMovie()
-    .then((movie) => {
-      renderMovies(movie);
-    })
-    .catch((err) => {
-      console.log(err);
-    });
-}
+// export default function setPopularMovie() {
+//   fetchPopularMovie()
+//     .then((movie) => {
+//       renderMovies(movie);
+//     })
+//     .catch((err) => {
+//       console.log(err);
+//     });
+// }
